+++ fabric/decorators.py
@@ -9,11 +9,29 @@ from fabric import tasks
 from .context_managers import settings
 
 
-def task(func):
+def task(*args, **kwargs):
     """
-    Decorator declaring the wrapped function as a :ref:`new-style task <new-style-tasks>`.
+    Decorator declaring the wrapped function to be a new-style task.
+
+    May be invoked as a simple, argument-less decorator (i.e. ``@task``) or
+    with arguments customizing its behavior (e.g. ``@task(alias='myalias')``).
+
+    Please see the :ref:`new-style task <task-decorator>` documentation for
+    details on how to use this decorator.
+
+    .. versionchanged:: 1.2
+        Added the ``alias``, ``aliases``, ``task_class`` and ``default``
+        keyword arguments. See :ref:`task-decorator-arguments` for details.
     """
-    return tasks.WrappedCallableTask(func)
+    invoked = bool(not args or kwargs)
+    task_class = kwargs.pop("task_class", tasks.WrappedCallableTask)
+    if not invoked:
+        func, args = args[0], ()
+
+    def wrapper(func):
+        return task_class(func, *args, **kwargs)
+
+    return wrapper if invoked else wrapper(func)
 
 
 def hosts(*host_list):