+++ fabric/main.py
@@ -163,11 +163,11 @@ def load_fabfile(path, importer=None):
         del sys.path[0]
 
     # Actually load tasks
-    docstring, new_style, classic = load_tasks_from_module(imported)
+    docstring, new_style, classic, default = load_tasks_from_module(imported)
     tasks = new_style if state.env.new_style_tasks else classic
     # Clean up after ourselves
     _seen.clear()
-    return docstring, tasks
+    return docstring, tasks, default
 
 
 def load_tasks_from_module(imported):
@@ -184,30 +184,48 @@ def load_tasks_from_module(imported):
     # Return a two-tuple value.  First is the documentation, second is a
     # dictionary of callables only (and don't include Fab operations or
     # underscored callables)
-    new_style, classic = extract_tasks(imported_vars)
-    return imported.__doc__, new_style, classic
+    new_style, classic, default = extract_tasks(imported_vars)
+    return imported.__doc__, new_style, classic, default
+
+
+# For attribute tomfoolery
+class _Dict(dict):
+    pass
 
 
 def extract_tasks(imported_vars):
     """
     Handle extracting tasks from a given list of variables
     """
-    new_style_tasks = defaultdict(dict)
+    new_style_tasks = _Dict()
     classic_tasks = {}
+    default_task = None
     if 'new_style_tasks' not in state.env:
         state.env.new_style_tasks = False
     for tup in imported_vars:
         name, obj = tup
         if is_task_object(obj):
             state.env.new_style_tasks = True
+            # Honor instance.name
             new_style_tasks[obj.name] = obj
+            # Handle aliasing
+            if obj.aliases is not None:
+                for alias in obj.aliases:
+                    new_style_tasks[alias] = obj
+            # Handle defaults
+            if obj.is_default:
+                default_task = obj
         elif is_classic_task(tup):
             classic_tasks[name] = obj
         elif is_task_module(obj):
-            docs, newstyle, classic = load_tasks_from_module(obj)
+            docs, newstyle, classic, default = load_tasks_from_module(obj)
             for task_name, task in newstyle.items():
+                if name not in new_style_tasks:
+                    new_style_tasks[name] = _Dict()
                 new_style_tasks[name][task_name] = task
-    return (new_style_tasks, classic_tasks)
+            if default is not None:
+                new_style_tasks[name].default = default
+    return new_style_tasks, classic_tasks, default_task
 
 
 def is_task_module(a):
@@ -332,6 +350,8 @@ def _task_names(mapping):
     tasks, collections = _sift_tasks(mapping)
     for collection in collections:
         module = mapping[collection]
+        if hasattr(module, 'default'):
+            tasks.append(collection)
         join = lambda x: ".".join((collection, x))
         tasks.extend(map(join, _task_names(module)))
     return tasks
@@ -348,7 +368,11 @@ def _crawl(name, mapping):
 
 def crawl(name, mapping):
     try:
-        return _crawl(name, mapping)
+        result = _crawl(name, mapping)
+        # Handle default tasks
+        if isinstance(result, _Dict) and getattr(result, 'default', False):
+            result = result.default
+        return result
     except (KeyError, TypeError):
         return None
 
@@ -635,21 +659,15 @@ def main():
             print("Fabric %s" % state.env.version)
             sys.exit(0)
 
-        # Handle case where we were called bare, i.e. just "fab", and print
-        # a help message.
-        actions = (options.list_commands, options.shortlist, options.display,
-            arguments, remainder_arguments)
-        if not any(actions):
-            parser.print_help()
-            sys.exit(1)
-
         # Load settings from user settings file, into shared env dict.
         state.env.update(load_settings(state.env.rcfile))
 
         # Find local fabfile path or abort
         fabfile = find_fabfile()
         if not fabfile and not remainder_arguments:
-            abort("Couldn't find any fabfiles!")
+            abort("""Couldn't find any fabfiles!
+
+Remember that -f can be used to specify fabfile path, and use -h for help.""")
 
         # Store absolute path to fabfile in case anyone needs it
         state.env.real_fabfile = fabfile
@@ -658,8 +676,16 @@ def main():
         # tweaks to env values) and put its commands in the shared commands
         # dict
         if fabfile:
-            docstring, callables = load_fabfile(fabfile)
+            docstring, callables, default = load_fabfile(fabfile)
             state.commands.update(callables)
+
+        # Handle case where we were called bare, i.e. just "fab", and print
+        # a help message.
+        actions = (options.list_commands, options.shortlist, options.display,
+            arguments, remainder_arguments, default)
+        if not any(actions):
+            parser.print_help()
+            sys.exit(1)
 
         # Abort if no commands found
         if not state.commands and not remainder_arguments:
@@ -687,7 +713,7 @@ def main():
             display_command(options.display)
 
         # If user didn't specify any commands to run, show help
-        if not (arguments or remainder_arguments):
+        if not (arguments or remainder_arguments or default):
             parser.print_help()
             sys.exit(0)  # Or should it exit with error (1)?
 
@@ -713,6 +739,10 @@ def main():
             r = '<remainder>'
             state.commands[r] = lambda: api.run(remainder_command)
             commands_to_run.append((r, [], {}, [], [], []))
+
+        # Ditto for a default, if found
+        if not commands_to_run and default:
+            commands_to_run.append((default.name, [], {}, [], [], []))
 
         if state.output.debug:
             names = ", ".join(x[0] for x in commands_to_run)