+++ fabric/tasks.py
@@ -15,8 +15,17 @@ class Task(object):
     """
     name = 'undefined'
     use_task_objects = True
+    aliases = None
+    is_default = False
 
     # TODO: make it so that this wraps other decorators as expected
+    def __init__(self, alias=None, aliases=None, default=False,
+        *args, **kwargs):
+        if alias is not None:
+            self.aliases = [alias, ]
+        if aliases is not None:
+            self.aliases = aliases
+        self.is_default = default
 
     def run(self):
         raise NotImplementedError
@@ -26,13 +35,12 @@ class WrappedCallableTask(Task):
     """
     Wraps a given callable transparently, while marking it as a valid Task.
 
-    Generally used via the `~fabric.decorators.task` decorator and not
-    directly.
+    Generally used via `@task <~fabric.decorators.task>` and not directly.
 
     .. versionadded:: 1.1
     """
-    def __init__(self, callable):
-        super(WrappedCallableTask, self).__init__()
+    def __init__(self, callable, *args, **kwargs):
+        super(WrappedCallableTask, self).__init__(*args, **kwargs)
         self.wrapped = callable
         self.__name__ = self.name = callable.__name__
         self.__doc__ = callable.__doc__