+++ tests/support/default_task_submodule.py
@@ -1,0 +1,5 @@
+from fabric.api import task
+
+@task(default=True)
+def long_task_name():
+    pass