+++ tests/support/flat_alias.py
@@ -1,0 +1,5 @@
+from fabric.api import task
+
+@task(alias="foo_aliased")
+def foo():
+    pass