+++ tests/support/flat_aliases.py
@@ -1,0 +1,5 @@
+from fabric.api import task
+
+@task(aliases=["foo_aliased", "foo_aliased_two"])
+def foo():
+    pass