+++ tests/test_decorators.py
@@ -1,4 +1,5 @@
 from nose.tools import eq_, ok_
+import fudge
 from fudge import Fake, with_fakes
 import random
 
@@ -55,3 +56,70 @@ def test_with_settings_passes_env_vars_into_decorated_function():
     ok_(some_task(), msg="sanity check")
     eq_(random_return, decorated_task())
 
+
+def test_will_invoked_whatever_class_you_provide():
+    def foo(): pass
+    fake = Fake()
+    fake.expects("__init__").with_args(foo)
+    fudge.clear_calls()
+    fudge.clear_expectations()
+
+    foo = decorators.task(foo, task_class=fake)
+
+    fudge.verify()
+
+
+def test_passes_args_to_the_task_class():
+    random_vars = ("some text", random.randint(100, 200))
+    def foo(): pass
+
+    fake = Fake()
+    fake.expects("__init__").with_args(foo, *random_vars)
+    fudge.clear_calls()
+    fudge.clear_expectations()
+
+    foo = decorators.task(foo, task_class=fake, *random_vars)
+    fudge.verify()
+
+
+def test_passes_kwargs_to_the_task_class():
+    random_vars = {
+        "msg": "some text",
+        "number": random.randint(100, 200),
+    }
+    def foo(): pass
+
+    fake = Fake()
+    fake.expects("__init__").with_args(foo, **random_vars)
+    fudge.clear_calls()
+    fudge.clear_expectations()
+
+    foo = decorators.task(foo, task_class=fake, **random_vars)
+    fudge.verify()
+
+
+def test_integration_tests_for_invoked_decorator_with_no_args():
+    r = random.randint(100, 200)
+    @decorators.task()
+    def foo():
+        return r
+
+    eq_(r, foo())
+
+
+def test_integration_tests_for_decorator():
+    r = random.randint(100, 200)
+    @decorators.task(task_class=tasks.WrappedCallableTask)
+    def foo():
+        return r
+
+    eq_(r, foo())
+
+
+def test_original_non_invoked_style_task():
+    r = random.randint(100, 200)
+    @decorators.task
+    def foo():
+        return r
+
+    eq_(r, foo())